Move public exports from `exports.py` to `backend/__init__py`.
The main motivation is to use the exported `keras.Variable` class throughout code.
The exported `keras.Variable` class is a subclass of the backend specific implementation. However, that public class was not used internally to create variables, for instance via `add_weight()`.
There was no convenient way to detect that a tensor really is a variable. This code would print `False`:
```python
v = self.add_weight(x)
print(isinstance(v, keras.Variable)
```

Additionally, made API exposed in `keras/src/backend/<backend>/__init__.py` consistent between each other and consistent with `keras.src.backend`, which is important when using the `DynamicBackend` feature:
- added `device_scope()` to Numpy backend (a no-op)
- added import for `name_scope()` in all backends
- fixed bug where `SeedGenerator` was not using the dynamic backend specific `name_scope`
