+++ keras/api/__init__.py
@@ -31,12 +31,12 @@ from keras.api import regularizers
 from keras.api import saving
 from keras.api import tree
 from keras.api import utils
+from keras.src.backend import Variable
+from keras.src.backend import device
+from keras.src.backend import name_scope
 from keras.src.backend.common.keras_tensor import KerasTensor
 from keras.src.backend.common.stateless_scope import StatelessScope
 from keras.src.backend.common.symbolic_scope import SymbolicScope
-from keras.src.backend.exports import Variable
-from keras.src.backend.exports import device
-from keras.src.backend.exports import name_scope
 from keras.src.dtype_policies.dtype_policy import DTypePolicy
 from keras.src.dtype_policies.dtype_policy import FloatDTypePolicy
 from keras.src.initializers.initializer import Initializer
