+++ keras/api/_tf_keras/keras/__init__.py
@@ -29,12 +29,12 @@ from keras.api._tf_keras.keras import layers
 from keras.api._tf_keras.keras import losses
 from keras.api._tf_keras.keras import metrics
 from keras.api._tf_keras.keras import preprocessing
+from keras.src.backend import Variable
+from keras.src.backend import device
+from keras.src.backend import name_scope
 from keras.src.backend.common.keras_tensor import KerasTensor
 from keras.src.backend.common.stateless_scope import StatelessScope
 from keras.src.backend.common.symbolic_scope import SymbolicScope
-from keras.src.backend.exports import Variable
-from keras.src.backend.exports import device
-from keras.src.backend.exports import name_scope
 from keras.src.dtype_policies.dtype_policy import DTypePolicy
 from keras.src.dtype_policies.dtype_policy import FloatDTypePolicy
 from keras.src.initializers.initializer import Initializer
