+++ keras/src/backend/numpy/__init__.py
@@ -1,3 +1,4 @@
+from keras.src.backend.common.name_scope import name_scope
 from keras.src.backend.numpy import core
 from keras.src.backend.numpy import image
 from keras.src.backend.numpy import linalg
@@ -12,6 +13,7 @@ from keras.src.backend.numpy.core import compute_output_spec
 from keras.src.backend.numpy.core import cond
 from keras.src.backend.numpy.core import convert_to_numpy
 from keras.src.backend.numpy.core import convert_to_tensor
+from keras.src.backend.numpy.core import device_scope
 from keras.src.backend.numpy.core import is_tensor
 from keras.src.backend.numpy.core import random_seed_dtype
 from keras.src.backend.numpy.core import shape
