+++ keras/src/backend/tests/device_scope_test.py
@@ -12,10 +12,10 @@ class DeviceTest(testing.TestCase):
         if not tf.config.list_physical_devices("GPU"):
             self.skipTest("Need at least one GPU for testing")
 
-        with backend.device_scope("cpu:0"):
+        with backend.device("cpu:0"):
             t = backend.numpy.ones((2, 1))
             self.assertIn("CPU:0", t.device)
-        with backend.device_scope("CPU:0"):
+        with backend.device("CPU:0"):
             t = backend.numpy.ones((2, 1))
             self.assertIn("CPU:0", t.device)
 
@@ -24,7 +24,7 @@ class DeviceTest(testing.TestCase):
         self.assertIn("GPU:0", t.device)
 
         # Also verify the explicit gpu device
-        with backend.device_scope("gpu:0"):
+        with backend.device("gpu:0"):
             t = backend.numpy.ones((2, 1))
             self.assertIn("GPU:0", t.device)
 
@@ -42,10 +42,10 @@ class DeviceTest(testing.TestCase):
         if platform != "gpu":
             self.skipTest("Need at least one GPU for testing")
 
-        with backend.device_scope("cpu:0"):
+        with backend.device("cpu:0"):
             t = backend.numpy.ones((2, 1))
             self.assertEqual(get_device(t), jax.devices("cpu")[0])
-        with backend.device_scope("CPU:0"):
+        with backend.device("CPU:0"):
             t = backend.numpy.ones((2, 1))
             self.assertEqual(get_device(t), jax.devices("cpu")[0])
 
@@ -54,14 +54,14 @@ class DeviceTest(testing.TestCase):
         self.assertEqual(get_device(t), jax.devices("gpu")[0])
 
         # Also verify the explicit gpu device
-        with backend.device_scope("gpu:0"):
+        with backend.device("gpu:0"):
             t = backend.numpy.ones((2, 1))
             self.assertEqual(get_device(t), jax.devices("gpu")[0])
 
     @pytest.mark.skipif(backend.backend() != "jax", reason="jax only")
     def test_invalid_jax_device(self):
         with self.assertRaisesRegex(ValueError, "Received: device_name='123'"):
-            backend.device_scope(123).__enter__()
+            backend.device(123).__enter__()
 
     @pytest.mark.skipif(backend.backend() != "torch", reason="torch only")
     def test_torch_device_scope(self):
@@ -70,10 +70,10 @@ class DeviceTest(testing.TestCase):
         if not torch.cuda.device_count():
             self.skipTest("Need at least one GPU for testing")
 
-        with backend.device_scope("cpu:0"):
+        with backend.device("cpu:0"):
             t = backend.numpy.ones((2, 1))
             self.assertEqual(t.device, torch.device("cpu"))
-        with backend.device_scope("CPU:0"):
+        with backend.device("CPU:0"):
             t = backend.numpy.ones((2, 1))
             self.assertEqual(t.device, torch.device("cpu"))
 
@@ -82,11 +82,11 @@ class DeviceTest(testing.TestCase):
         self.assertEqual(t.device, torch.device("cuda", 0))
 
         # Also verify the explicit gpu -> cuda conversion
-        with backend.device_scope("gpu:0"):
+        with backend.device("gpu:0"):
             t = backend.numpy.ones((2, 1))
             self.assertEqual(t.device, torch.device("cuda", 0))
 
     @pytest.mark.skipif(backend.backend() != "torch", reason="torch only")
     def test_invalid_torch_device(self):
         with self.assertRaisesRegex(ValueError, "Received: device_name='123'"):
-            backend.device_scope(123).__enter__()
+            backend.device(123).__enter__()
