+++ keras/src/layers/preprocessing/normalization_test.py
@@ -96,12 +96,10 @@ class NormalizationTest(testing.TestCase):
         reason="Test symbolic call for torch meta device.",
     )
     def test_call_on_meta_device_after_built(self):
-        from keras.src.backend.torch import core
-
         layer = layers.Normalization()
         data = np.random.random((32, 4))
         layer.adapt(data)
-        with core.device_scope("meta"):
+        with backend.device("meta"):
             layer(data)
 
     def test_normalization_with_mean_only_raises_error(self):
