+++ keras/src/utils/torch_utils.py
@@ -2,6 +2,7 @@ import io
 
 from packaging.version import parse
 
+from keras.src import backend
 from keras.src.api_export import keras_export
 from keras.src.layers import Layer
 from keras.src.ops import convert_to_numpy
@@ -101,12 +102,10 @@ class TorchModuleWrapper(Layer):
         return self.module.parameters(recurse=recurse)
 
     def _track_module_parameters(self):
-        from keras.src.backend.torch import Variable
-
         for param in self.module.parameters():
             # The Variable will reuse the raw `param`
             # and simply wrap it.
-            variable = Variable(
+            variable = backend.Variable(
                 initializer=param, trainable=param.requires_grad
             )
             self._track_variable(variable)
